Final
Problem Description:
Write a program that receives a graph as input and determines the connectivity
of the graph and the number of the connected components in it (the graph is
undirected).
Figure 1 A disconnected graph with three components
A graph is connected when there is a path between every pair of vertices. A graph
G is said to be disconnected if there exist at least two nodes in G such that no
path in G has those nodes as endpoints. A connected component (or just
component) of a graph is a sub graph in which any two vertices are connected
to each other by paths, and which is connected to no additional vertices in the
super graph.